feat: boundaries
