+++ backend/Core/Dto/AccountDto.cs
@@ -1,0 +1,3 @@
+﻿namespace Core.Dto;
+
+public sealed record AccountDto(string Name, decimal Balance, decimal TotalIncome, decimal TotalExpense, string Email);
