+++ backend/Core/UseCase/GetAccount/Boundaries/GetAccountInput.cs
@@ -1,0 +1,6 @@
+﻿using Core.Commons;
+using MediatR;
+
+namespace Core.UseCase.GetAccount.Boundaries;
+
+public sealed record GetAccountInput(Guid Id) : IRequest<Output>;
